--- a/.github/workflows/ci.yaml
+++ b/.github/workflows/ci.yaml
@@ -12,12 +12,24 @@ jobs:
   build-cabal:
     name: CI-cabal
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          # Linux
+          - { cabal: "3.4", os: ubuntu-latest,  ghc: "7.10.3" }
+          - { cabal: "3.4", os: ubuntu-latest,  ghc: "8.0.2"  }
+          - { cabal: "3.4", os: ubuntu-latest,  ghc: "8.2.2"  }
+          - { cabal: "3.4", os: ubuntu-latest,  ghc: "latest" }
+          # MacOS
+          - { cabal: "3.4", os: macOS-latest,   ghc: "latest" }
     steps:
     - uses: actions/checkout@v2
     - uses: actions/setup-haskell@v1
       id: setup-haskell-cabal
       with:
-        ghc-version: 'latest'
+        ghc-version: ${{ matrix.ghc }}
+        cabal-version: ${{ matrix.cabal }}
     - name: Update cabal package database
       run: cabal update
     - uses: actions/cache@v2
@@ -32,7 +44,12 @@ jobs:
         cabal sdist -z -o .
         cabal get random-*.tar.gz
         cd random-*/
-        cabal test all
+        if [ ${{matrix.ghc}} = 7.10.3 ]; then
+          EXTRA_FLAGS=--ghc-option="-optl-no-pie"
+        fi
+        echo $EXTRA_FLAGS
+        cabal $EXTRA_FLAGS configure --haddock-all --enable-tests --enable-benchmarks --benchmark-option=-l
+        cabal $EXTRA_FLAGS build all --write-ghc-environment-files=always
   build-stack:
     name: CI-stack
     runs-on: ${{ matrix.os }}
@@ -40,17 +57,8 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest]
-        resolver: [nightly, lts-17, lts-16, lts-14, lts-12, lts-11, lts-9, lts-6]
+        resolver: [nightly, lts-18, lts-16, lts-14, lts-12]
         include:
-          - resolver: lts-6
-            ghc: 7.10.3
-            stack-yaml: stack-really-old.yaml
-          - resolver: lts-9
-            ghc: 8.0.2
-            stack-yaml: stack-really-old.yaml
-          - resolver: lts-11
-            ghc: 8.2.2
-            stack-yaml: stack-old.yaml
           - resolver: lts-12
             ghc: 8.4.4
             stack-yaml: stack-old.yaml
@@ -58,12 +66,12 @@ jobs:
             stack-yaml: stack-coveralls.yaml
           - resolver: lts-16
             ghc: 8.8.4
-            stack-yaml: stack.yaml
-          - resolver: lts-17
+            stack-yaml: stack-old.yaml
+          - resolver: lts-18
             ghc: 8.10.4
             stack-yaml: stack.yaml
           - resolver: nightly
-            stack-yaml: stack-nightly.yaml
+            stack-yaml: stack.yaml
           # Latest stable for MacOS: ghc-8.8.4
           - resolver: lts-16
             os: macos-latest