--- a/stack.yaml
+++ b/stack.yaml
@@ -1,4 +1,4 @@
-resolver: lts-16.31
+resolver: lts-18.0
 packages:
 - .
 extra-deps: