Benchmarks for `floating` point numbers seem suspicious, as it was rightly pointed out in #112 

This PR is designed to fix that issue by relying on writing generated random values into a `MutablePrimArray`, thus avoiding any potential ghc optimizations that can discard intermediate computations.

Before this PR:

```
      floating      
        IO          
          uniformFloat01M:          OK (0.80s)
            3.1 ms ± 194 μs
          uniformFloatPositive01M:  OK (0.19s)
            3.1 ms ± 260 μs
          uniformDouble01M:         OK (0.24s)
             58 μs ± 4.5 μs
          uniformDoublePositive01M: OK (0.18s)
             86 μs ± 6.3 μs
        pure        
          uniformFloat01M:          OK (0.16s)
            334 μs ±  25 μs
          uniformFloatPositive01M:  OK (0.19s)
            351 μs ±  26 μs
          uniformDouble01M:         OK (0.20s)
            760 μs ±  60 μs
          uniformDoublePositive01M: OK (0.21s)
            768 μs ±  54 μs
```

With this PR (only `floating/IO` has changed):

```
      floating
        IO
          uniformFloat01M:          OK (0.19s)
            370 μs ±  22 μs
          uniformFloatPositive01M:  OK (0.20s)
            382 μs ±  24 μs
          uniformDouble01M:         OK (0.20s)
            772 μs ±  55 μs
          uniformDoublePositive01M: OK (0.21s)
            801 μs ±  51 μs
        pure
          uniformFloat01M:          OK (0.17s)
            353 μs ±  24 μs
          uniformFloatPositive01M:  OK (0.76s)
            371 μs ±  19 μs
          uniformDouble01M:         OK (0.19s)
            750 μs ±  57 μs
          uniformDoublePositive01M: OK (0.19s)
            749 μs ±  72 μs
```

This PR also changes CI so older ghc versions are tested with cabal instead of stack, because new dependency `primitive-0.7.1.0` doesn't work with older Cabal the library, which is deeply ingrained into each LTS.