@andrewthad 2021-06-30
Good work! I often use a similar approach in my own benchmarking, and it works well. The only thing to be careful of is that you really want to be sure that fillMutablePrimArrayM is inlined. The best way to do this is to annotate `fillMutablePrimArrayM` with `{-# inline fillMutablePrimArrayM #-}` and then to hoist all your tests into top-level non-inlined functions, turning this:

    , env ((,) <$> getStdGen <*> newAlignedPinnedPrimArray sz) $ \ ~(gen, ma) ->
        bench "uniformDouble01M" $
        nfIO (runStateGenT gen (fillMutablePrimArrayM uniformDouble01M ma))

Into this:

    , env ((,) <$> getStdGen <*> newAlignedPinnedPrimArray sz) $ \ ~(gen, ma) ->
        bench "uniformDouble01M" $
        nfIO (runStateGenT gen uniformDouble01M_specialized))
    ...
    uniformDouble01M_specialized :: StdGen -> ST s (PrimArray Double)
    {-# noinline uniformDouble01M_specialized #-}
    uniformDouble01M_specialized gen = fillMutablePrimArrayM uniformDouble01M ma gen

This makes the benchmarks more verbose, but it has several advantages for these kinds of low-level benchmarks:

1. The specialization is more reliable.
2. You can use `-ddump-simpl` and find `uniformDouble01M_specialized` and look for perf issues. If it's inlined into the massive blob of benchmark suite stuff, the dumped Core is difficult to find stuff in.
3. This prevents issues with code alignment, which can make a big difference in benchmarks that loops with short bodies. Read Andreas's response to my issue at https://gitlab.haskell.org/ghc/ghc/-/issues/16556 for a better explanation of this.