@lehins 2021-06-30
Thank you @andrewthad for the review!

> The only thing to be careful of is that you really want to be sure that fillMutablePrimArrayM is inlined.

Yeah, I am quite familiar with this problem :) However, it is usually quite obvious from the benchmark results when a function is not inlined, its performance becomes terrible. Also, from what I know, when a function is defined in the same module, that it is being benchmarked in (i.e. it is where main is and it is not exported), then ghc will almost always inline it. So, I think `INLINE` pragma is is more of a concern when a function is imported from some other module.

>  hoist all your tests into top-level non-inlined functions

That is certainly an interesting suggestion, which I'll definitely have to take into consideration and experiment with. All the benefits you outlined make a lot of sense. However, I'll refrain from adding this approach in this PR, since, as far as I understand it, there should be little affect on the performance results. Most importantly though I have too much on my plate as it is at the moment to mess with this. I am pretty confident that the actual implementation of floating point number generation is pretty good, so I am not worried too much about those benchmarks.