@andrewthad 2021-06-30
No worries. Just something to keep in mind if someone ever reports wobbly results across different versions of GHC.